Refactored TableauServer for better modularity
