+++ tableau_utilities/__init__.py
@@ -1,5 +1,6 @@
 from .scripts import cli
 from .tableau_file.tableau_file import TableauFileError, Datasource
 from .tableau_file import tableau_file_objects
-from .tableau_server.tableau_server import TableauConnectionError, TableauServer
+from .tableau_server.static import TableauConnectionError
+from .tableau_server.tableau_server import TableauServer
 from .tableau_server import tableau_server_objects
